Refactor 募集のフォームを再利用可能に
